change ids from int to long

+++ src/main/java/com/backend/theWizardsBag/models/Item.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class Item {
 
     // ATTs
-    private int itemId;
+    private long itemId;
     private String name;
     private String ttrpg;
     private boolean magical;                //   No Null
