+++ src/main/java/com/backend/theWizardsBag/models/Note.java
@@ -3,7 +3,7 @@ package com.backend.theWizardsBag.models;
 public class Note {
 
     // ATTs
-    private int noteId;
+    private long noteId;
     private String date;
     private String note;
     private int authorId;       //   User Id
